+++ js/data.js
@@ -1,41 +1,5 @@
 import {getRandomNumbarValue} from './util.js';
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const NAMES = [
-  'Андрей',
-  'Алена',
-  'Ольга',
-  'Лев',
-  'Максим',
-  'Игорь',
-  'Владимир',
-  'Елена',
-  'Рома',
-  'Сергей',
-];
-
-const DESCRIPTION = [
-  'Phasellus quis mauris ac nibh.',
-  'Duis pellentesque tincidunt urna vel.',
-  'Vestibulum ante ipsum primis in.',
-  'Curabitur laoreet tristique luctus. Quisque.',
-  'Integer vitae augue in ligula.',
-  'Quisque aliquam purus ac dui.',
-  'Morbi iaculis imperdiet pretium. Proin.',
-  'Integer ullamcorper, eros in porttitor.',
-  'Donec pulvinar metus est, et.',
-  'Sed ac velit quis neque.',
-];
-
-const NECESSARY_SUM_OBJECTS = 25;
+import {MESSAGES, NAMES, DESCRIPTION} from './const.js';
 
 const generateData = (num, functionGenerate) => {
   const arrayObjects = [];
@@ -72,5 +36,4 @@ const generateObject = (num) => {
   return object;
 };
 
-// Массив нужных данных
-generateData(NECESSARY_SUM_OBJECTS, generateObject);
+export {generateData, generateObject};
